pyproject.toml
[tool.poetry]
name = "grist-uffd"
version = "0.1.0"
description = ""
authors = ["Russ Garrett <russ@garrett.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4
